+++ R/code2pptx.R
@@ -19,7 +19,7 @@
 #' @importFrom devEMF emf
 #' @export
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' code2office(plot(iris))
 #' require(ggplot2)
 #' gg=ggplot(data=mtcars,aes(x=wt,y=mpg))+geom_point()
@@ -33,6 +33,11 @@ code2office=function(...,ggobj=NULL,target="Report",append=FALSE,title="",
     # type="pptx";preprocessing="";plottype="auto";echo=FALSE;parallel=FALSE
     # left=1;top=1;width=NULL;height=NULL;aspectr=NULL
 
+    if(preprocessing!=""){
+        #sink("NUL")
+        eval(parse(text=preprocessing),envir = global_env())
+        #unsink("NUL")
+    }
     if(is.null(width)){
         if(is.null(height)){
             if(is.null(aspectr)){
@@ -113,7 +118,7 @@ code2office=function(...,ggobj=NULL,target="Report",append=FALSE,title="",
 #' @param ... further arguments to be passed to code2office
 #' @export
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' code2pptx(plot(iris))
 #' require(ggplot2)
 #' gg=ggplot(data=mtcars,aes(x=wt,y=mpg))+geom_point()
@@ -128,7 +133,7 @@ code2pptx=function(...){
 #' @param ... further arguments to be passed to code2office
 #' @export
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' code2docx(plot(iris))
 #' require(ggplot2)
 #' gg=ggplot(data=mtcars,aes(x=wt,y=mpg))+geom_point()