+++ R/table2office.R
@@ -5,16 +5,23 @@
 #' @param title Optional character of plot title
 #' @param vanilla A logical
 #' @param echo logical
+#' @param add.rownames logical
 #' @param preprocessing A character string
 #' @param type "pptx" or "docx"
 #' @param landscape logical
 #' @param left left margin
 #' @param top top margin
 #' @importFrom moonBook mytable
+#' @importFrom flextable autofit
 #' @export
-table2office=function(x=NULL,target="Report",append=FALSE,title="",vanilla=FALSE,echo=FALSE,
+table2office=function(x=NULL,target="Report",append=FALSE,title="",vanilla=FALSE,echo=FALSE,add.rownames=TRUE,
                       preprocessing="",type="pptx",landscape=FALSE,left=1,top=1){
 
+    if(preprocessing!=""){
+        #sink("NUL")
+        eval(parse(text=preprocessing),envir = global_env())
+        #unsink("NUL")
+    }
     doc<-open_doc(target=target,type=type,append=append)
     target=attr(doc,"name")
     if(title!=""){
@@ -27,7 +34,7 @@ table2office=function(x=NULL,target="Report",append=FALSE,title="",vanilla=FALSE
     if(title!="") pos=pos+0.5
     if(echo & is.character(x)) {
 
-        codeft=Rcode2flextable(x,preprocessing=preprocessing,format="pptx",eval=FALSE)
+        codeft=Rcode2flextable(x,format="pptx",eval=FALSE)
         doc<-doc %>% ph_with(value=codeft, location = ph_location(left=1,top=pos))
         pos=pos+0.5
 
@@ -40,13 +47,13 @@ table2office=function(x=NULL,target="Report",append=FALSE,title="",vanilla=FALSE
         ft<-mytable2flextable(x,vanilla=vanilla)
     } else if(class(x)[1] %in%
               c("matrix","lm","fitdistr","nls","aov","anova","glm","coxph","prcomp","summary.prcomp")){
-        ft<-ztable2flextable(ztable(x),vanilla=vanilla)
-    } else if(class(x)[1]=="data.frame"){
-        ft<-df2flextable(x,vanilla=vanilla)
-    } else if(class(x)[1]=="flextable"){
+        ft<-autofit(ztable2flextable(ztable(x),add.rownames=add.rownames,vanilla=vanilla))
+    }  else if(class(x)[1]=="flextable"){
         ft<-x
+    } else if("data.frame" %in% class(x)){
+        ft<-df2flextable(x,vanilla=vanilla,add.rownames=add.rownames)
     }
-    if(class(doc)=="rpptx"){
+    if(inherits(doc,"rpptx")){
         doc<-doc %>% ph_with(value=ft,location = ph_location(left=left,top=pos))
     } else {
         if(landscape) doc <- body_end_section_portrait(doc)
@@ -61,7 +68,7 @@ table2office=function(x=NULL,target="Report",append=FALSE,title="",vanilla=FALSE
 #' @param ... further arguments to be passed to table2office
 #' @export
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' require(moonBook)
 #' x="mytable(Dx~.,data=acs)"
 #' table2pptx(x,title="mytable object",echo=TRUE)
@@ -80,7 +87,7 @@ table2pptx=function(...){
 #' @param ... further arguments to be passed to table2office
 #' @export
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' require(moonBook)
 #' x=mytable(Dx~.,data=acs)
 #' table2docx(x)